+++ crates/telio-model/src/api_config.rs
@@ -84,6 +84,8 @@ pub struct FeatureNurse {
     pub fingerprint: String,
     /// Heartbeat interval in seconds. Default value is 3600.
     pub heartbeat_interval: Option<u32>,
+    /// Initial heartbeat interval in seconds. Default value is None.
+    pub initial_heartbeat_interval: Option<u32>,
     /// QoS configuration for Nurse
     pub qos: Option<FeatureQoS>,
     /// Enable/disable collecting nat type
@@ -316,6 +318,7 @@ mod tests {
         nurse: Some(FeatureNurse {
             fingerprint: "fingerprint_test".to_string(),
             heartbeat_interval: None,
+            initial_heartbeat_interval: None,
             qos: None,
             enable_nat_type_collection: None,
         }),
@@ -482,6 +485,7 @@ mod tests {
             nurse: Some(FeatureNurse {
                 fingerprint: String::from("fingerprint_test"),
                 heartbeat_interval: Some(3600),
+                initial_heartbeat_interval: None,
                 qos: Some(FeatureQoS {
                     rtt_interval: Some(3600),
                     rtt_tries: Some(5),
@@ -502,6 +506,7 @@ mod tests {
             nurse: Some(FeatureNurse {
                 fingerprint: String::from("fingerprint_test"),
                 heartbeat_interval: None,
+                initial_heartbeat_interval: None,
                 qos: Some(FeatureQoS {
                     rtt_interval: None,
                     rtt_tries: None,
@@ -522,6 +527,7 @@ mod tests {
             nurse: Some(FeatureNurse {
                 fingerprint: String::from("fingerprint_test"),
                 heartbeat_interval: None,
+                initial_heartbeat_interval: None,
                 qos: None,
                 enable_nat_type_collection: None,
             }),